third_party/WebKit/LayoutTests/http/tests/devtools/elements/styles/up-down-numerics-and-colors-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that numeric and color values are incremented/decremented correctly.


Running: testInit

Running: testAlterColor
    color: #EE3;

Running: testAlterNumber
    opacity: 11.6;

Running: testAlterBigNumber
    transform: rotate(1000000000000000065537deg);
